requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
polars>=0.20.0
pyarrow>=14.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
tensorboard>=2.14.0
PyYAML>=6.0
